requirements.txt
click==8.1.3
cycler==0.11.0
fonttools==4.33.3
h5py==3.7.0
kiwisolver==1.4.2
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
pyparsing==3.0.9
PyQRCode==1.2.1
PyQt5==5.15.6
pyqt5-plugins==5.15.4.2.2
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1
python-dateutil==2.8.2
PyYAML==6.0
qt5-applications==5.15.2.2.2
qt5-tools==5.15.2.1.2
six==1.16.0
traitlets==5.2.2.post1
